Types work in the built version again, you can actually use this now!
